Refactors transition
